Read a metadata yaml file and convert to atom4R DCMI object.

Metadata templates can be generated with deposits_metadata_template, and values manually entered. This functions loads the completed template and converts it to a DCEntry object from the atom4R package.

deposits_meta_to_dcmi(filename = NULL, id = "my-id")

Arguments

filename

Name of completed metadata file generated from deposits_metadata_template.

id

A unique identifier for the deposit, as a single character string.

See also

Examples

filename <- tempfile (fileext = ".json")
deposits_metadata_template (filename)
#> Loading Atom XML schemas...
#> Loading DCMI vocabularies...
m <- readLines (filename)
m [grep ("Title", m)] <- "  \"Title\": \"New Title\""
m [grep ("Type", m)] <- "  \"Type\": \"Software\""
m [grep ("Description", m)] <-
   "  \"Description\": \"Description of software\""
# Values can be entered in JSON format:
m [grep ("TableOfContents", m)] <-
    "  \"TableOfContents\": {\"one\": \"First\", \"two\": \"Second\"}"
dc <- deposits_meta_to_dcmi (filename, id = "my-id")
#> [atom4R][INFO] Object 'DCEntry' is VALID according to Atom XML schemas!